pyproject.toml
[tool.poetry]
name = "twitter-crypto-post"
version = "0.1.0"
description = ""
authors = ["Vebnik(Nik) <vebnik111@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.scripts]
tw = "main:tw_main"
bin = "main:bin_main"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.0"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^23.1a1", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
